Switch to coverage.py from pytest-cov
While pytest-cov was a little bit more ergonomic, it wasn't really
necessary, considering coverage.py was a nested dependency anyway. And
the way I was using the annotated format was wrong for files in
directories, and also deprecated. Swap to direct usage of coverage.py
and JSON, a very easy-to-parse format.